refactor(collie-core): group the IPC command handlers into modules
collie_core/ipc/server.py was 3,915 lines: connection handling, the frame
dispatch, shared helpers and all 135 _cmd_<kind> handlers in one class.
Working on the routine handlers meant reading past collaboration, providers,
connectors, files and the chat turn first.

93 handlers now live in collie_core/ipc/commands/, one module per command
family: collaboration (14), memory (17), routines (16), providers (17),
connectors (12), files (5), agents (8), messengers (4). They are mixed into
CollieIPCServer, which keeps the connection, the dispatch and the shared
server state, so nothing on the wire changes: every handler is still
server._cmd_<kind>, and dispatch was already indirect through
getattr(self, f"_cmd_{kind}") over the MRO.

Three names had one reader on each side of the new seam, so they moved to the
leaf module ipc/command_support.py (_MAX_LIST_LIMIT, _bounded_list_limit).
_PERSON_FIELDS moved to the memory commands module and _OAuthAttemptState to
the providers commands module, each of which was its only reader.

tests/collie/test_ipc_contract.py grepped handler names out of server.py alone,
so it now scans the whole ipc package: a handler moving between modules can no
longer change the wire contract by itself. tests/collie/test_ipc_command_split.py
is new and pins which module owns which handler, rejects a handler owned twice
(a silent MRO shadow), rejects a handler left behind in server.py, and requires
each module to declare the server state it reaches through self.

Verified: all 170 members of CollieIPCServer (135 handlers + 35 helpers)
compared with inspect.getsource between origin/main and this branch, 0 missing
and 0 changed; CI-exact backend run 3,757 passed / 2 skipped / 1 deselected at
75.71% coverage, exit 0 (main baseline 3,738, delta is the 19 new guard tests);
ruff check and format gates clean.

server.py is 2,456 lines now. The chat turn, the plan/run/approval flow,
conversation CRUD and data management stay: they share the chat task plumbing
and the run state, so they are one cluster rather than several families.

docs/engineering/architecture/ipc-command-modules.md records the seam and the
steps for adding a command; PROJECT_MAP.md and the repository snapshot updated.
